authtoken route moved to management API
